update: rename onMemberRemove to onMemberKick
