+++ src/events/onMemberKick.ts
